Deleted module aggregate.
