--- src/dcegm/aggregate_policy_value.py
@@ -1,23 +1,0 @@
-from typing import Callable
-
-import numpy as np
-
-
-def calc_value(
-    consumption: np.ndarray,
-    next_period_value: np.ndarray,
-    choice: int,
-    discount_factor: float,
-    compute_utility: Callable,
-) -> np.ndarray:
-    """Compute the agent's value in the credit constrained region.
-
-    Args:
-        compute_utility (callable): User-defined function to compute the agent's
-            utility. The input ``params``` is already partialled in.
-
-    """
-    utility = compute_utility(consumption, choice)
-    value_constrained = utility + discount_factor * next_period_value
-
-    return value_constrained
