+++ src/dcegm/egm_step.py
@@ -449,3 +449,23 @@ def calc_choice_probability(
     )
 
     return choice_prob
+
+
+def calc_value(
+    consumption: np.ndarray,
+    next_period_value: np.ndarray,
+    choice: int,
+    discount_factor: float,
+    compute_utility: Callable,
+) -> np.ndarray:
+    """Compute the agent's value in the credit constrained region.
+
+    Args:
+        compute_utility (callable): User-defined function to compute the agent's
+            utility. The input ``params``` is already partialled in.
+
+    """
+    utility = compute_utility(consumption, choice)
+    value_constrained = utility + discount_factor * next_period_value
+
+    return value_constrained
